Function for prediction at new locations for multi-species hierarchical distance sampling models

The function predict collects posterior predictive samples for a set of new locations given an object of class `msDS`. Prediction is possible for both the latent abundance state as well as detection.

Usage

# S3 method for msDS
predict(object, X.0, ignore.RE = FALSE, type = 'abundance', ...)

Arguments

object

an object of class msDS

X.0

the design matrix of covariates at the prediction locations. This should include a column of 1s for the intercept if an intercept is included in the model. If random effects are included in the abundance (or detection if type = 'detection') portion of the model, the levels of the random effects at the new locations should be included as a column in the design matrix. The ordering of the levels should match the ordering used to fit the data in msDS. Columns should correspond to the order of how covariates were specified in the corresponding formula argument of msDS. Column names must match the names of the variables used to fit the model (for the intercept, use '(Intercept)').

ignore.RE

a logical value indicating whether to include unstructured random effects for prediction. If TRUE, random effects will be ignored and prediction will only use the fixed effects. If FALSE, random effects will be included in the prediction for both observed and unobserved levels of the random effect.

type

a quoted keyword indicating what type of prediction to produce. Valid keywords are 'abundance' to predict expected abundance and latent abundance values (this is the default), or 'detection' to predict detection probability given new values of detection covariates.

...

currently no additional arguments

Note

When ignore.RE = FALSE, both sampled levels and non-sampled levels of random effects are supported for prediction. For sampled levels, the posterior distribution for the random effect corresponding to that level of the random effect will be used in the prediction. For non-sampled levels, random values are drawn from a normal distribution using the posterior samples of the random effect variance, which results in fully propagated uncertainty in predictions with models that incorporate random effects.

Author

Jeffrey W. Doser doserjef@msu.edu,

Value

A list object of class predict.msDS. When type = 'abundance', the list consists of:

mu.0.samples

a three-dimensional array of posterior predictive samples for the expected abundance values, or expected abundance values per unit area (i.e., density) values when an offset was used when fitting the model with msDS().

N.0.samples

a three-dimensional array of posterior predictive samples for the latent abundance values. These will be in the same units as mu.0.samples.

When type = 'detection', the list consists of:

sigma.0.samples

a three-dimensional array of posterior predictive samples for sigma (the parameter controlling detection probability).

The return object will include additional objects used for standard extractor functions.

Examples

set.seed(210)
J.x <- 10
J.y <- 10 
J <- J.x * J.y
# Number of distance bins from which to simulate data. 
n.bins <- 5
# Length of each bin. This should be of length n.bins
bin.width <- c(.10, .10, .20, .3, .1)
# Number of species
n.sp <- 5
# Community-level abundance coefficients
beta.mean <- c(-1, 0.2, 0.3, -0.2)
p.abund <- length(beta.mean)
tau.sq.beta <- c(0.2, 0.3, 0.5, 0.4)
# Detection coefficients
alpha.mean <- c(-1.0, -0.3)
p.det <- length(alpha.mean)
tau.sq.alpha <- c(0.1, 0.2)
# Detection decay function
det.func <- 'halfnormal'
mu.RE <- list()
p.RE <- list()
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = n.sp, ncol = p.abund)
alpha <- matrix(NA, nrow = n.sp, ncol = p.det)
for (i in 1:p.abund) {
  beta[, i] <- rnorm(n.sp, beta.mean[i], sqrt(tau.sq.beta[i]))
}
for (i in 1:p.det) {
  alpha[, i] <- rnorm(n.sp, alpha.mean[i], sqrt(tau.sq.alpha[i]))
}
sp <- FALSE 
family <- 'Poisson'
kappa <- runif(n.sp, 0.3, 3) 
offset <- pi * .8^2
transect <- 'line'
factor.model <- FALSE

dat <- simMsDS(J.x = J.x, J.y = J.y, n.bins = n.bins, bin.width = bin.width,
               n.sp = n.sp, beta = beta, alpha = alpha, det.func = det.func, kappa = kappa, 
               mu.RE = mu.RE, p.RE = p.RE, sp = sp, cov.model = cov.model,
               sigma.sq = sigma.sq, phi = phi, nu = nu, family = family, 
               offset = offset, transect = transect, factor.model = factor.model)
#> overdispersion parameter (kappa) is ignored when family == 'Poisson'

# Split into fitting and prediction data set
pred.indx <- sample(1:J, round(J * .25), replace = FALSE)
y <- dat$y[, -pred.indx, ]
# Occupancy covariates
X <- dat$X[-pred.indx, ]
# Prediction covariates
X.0 <- dat$X[pred.indx, ]
# Detection covariates
X.p <- dat$X.p[-pred.indx, , drop = FALSE]
X.p.0 <- dat$X.p[pred.indx, , drop = FALSE]
coords <- as.matrix(dat$coords[-pred.indx, ])
coords.0 <- as.matrix(dat$coords[pred.indx, ])
dist.breaks <- dat$dist.breaks

covs <- cbind(X, X.p)
colnames(covs) <- c('int.abund', 'abund.cov.1', 'abund.cov.2', 'abund.cov.3', 
                    'int.det', 'det.cov.1')

data.list <- list(y = y, 
                  covs = covs,
                  dist.breaks = dist.breaks, 
                  offset = offset)

# Priors
prior.list <- list(beta.comm.normal = list(mean = 0, var = 10),
       alpha.comm.normal = list(mean = 0,
                     var = 10), 
                   kappa.unif = list(0, 100), 
                   tau.sq.beta.ig = list(a = 0.1, b = 0.1),
                   tau.sq.alpha.ig = list(a = 0.1, b = 0.1)) 
# Starting values
inits.list <- list(alpha.comm = 0, beta.comm = 0, beta = 0,
       alpha = 0, kappa = 1)

tuning <- list(beta = 0.1, alpha = 0.1, beta.star = 0.3, alpha.star = 0.1, 
               kappa = 0.8) 

n.batch <- 4 
batch.length <- 25
n.burn <- 0
n.thin <- 1
n.chains <- 1

out <- msDS(abund.formula = ~ abund.cov.1 + abund.cov.2 + abund.cov.3,
            det.formula = ~ det.cov.1,
            data = data.list, 
            n.batch = n.batch, 
            batch.length = batch.length, 
            inits = inits.list, 
            family = 'Poisson',
            det.func = 'halfnormal', 
            transect = transect, 
            tuning = tuning,
            priors = prior.list, 
            accept.rate = 0.43, 
            n.omp.threads = 1, 
            verbose = TRUE, 
            n.report = 10,
            n.burn = n.burn,
            n.thin = n.thin,
            n.chains = n.chains) 
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> N is not specified in initial values.
#> Setting initial values based on observed data
#> tau.sq.beta is not specified in initial values.
#> Setting initial values to random values between 0.05 and 1
#> tau.sq.alpha is not specified in initial values.
#> Setting to initial values to random values between 0.05 and 1
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> Multi-species Poisson HDS model with 75 sites and 5 species.
#> 
#> Samples per Chain: 100 (4 batches of length 25)
#> Burn-in: 0 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 100 
#> 
#> Source compiled with OpenMP support and model fit using 1 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Batch: 4 of 4, 100.00%
summary(out, level = 'community')
#> 
#> Call:
#> msDS(abund.formula = ~abund.cov.1 + abund.cov.2 + abund.cov.3, 
#>     det.formula = ~det.cov.1, data = data.list, inits = inits.list, 
#>     priors = prior.list, tuning = tuning, n.batch = n.batch, 
#>     batch.length = batch.length, accept.rate = 0.43, family = "Poisson", 
#>     transect = transect, det.func = "halfnormal", n.omp.threads = 1, 
#>     verbose = TRUE, n.report = 10, n.burn = n.burn, n.thin = n.thin, 
#>     n.chains = n.chains)
#> 
#> Samples per Chain: 100
#> Burn-in: 0
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 100
#> Run Time (min): 0.0039
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Abundance Means (log scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept) -0.7436 0.4285 -1.5066 -0.7859 0.0448   NA   9
#> abund.cov.1  0.3172 0.3651 -0.4399  0.3043 1.0424   NA 100
#> abund.cov.2  0.4615 0.3761 -0.2207  0.5081 1.1315   NA 100
#> abund.cov.3 -0.0533 0.3353 -0.6963 -0.0788 0.7151   NA 231
#> 
#> Abundance Variances (log scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.4657 1.0352 0.0671 0.2157 2.2611   NA  35
#> abund.cov.1 0.6553 1.0724 0.0749 0.4262 2.0974   NA  59
#> abund.cov.2 0.6086 0.6170 0.0612 0.3987 2.0154   NA  44
#> abund.cov.3 0.5607 0.5083 0.0506 0.4378 2.0249   NA  11
#> 
#> Detection Means (log scale): 
#>                Mean     SD    2.5%     50%  97.5% Rhat ESS
#> (Intercept) -0.5745 0.4009 -1.1166 -0.6719 0.1364   NA   3
#> det.cov.1   -0.3663 0.3023 -1.0171 -0.3883 0.1791   NA  12
#> 
#> Detection Variances (log scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.2665 0.2830 0.0328 0.1925 1.1759   NA  69
#> det.cov.1   0.3051 0.3612 0.0394 0.1937 0.9873   NA  38

# Predict at new locations ------------------------------------------------
colnames(X.0) <- c('intercept', 'abund.cov.1', 'abund.cov.2', 'abund.cov.3')
out.pred <- predict(out, X.0)
str(out.pred)
#> List of 3
#>  $ mu.0.samples: num [1:100, 1:5, 1:25] 0.961 0.894 0.6 0.522 0.491 ...
#>  $ N.0.samples : int [1:100, 1:5, 1:25] 0 5 1 0 0 0 0 0 1 0 ...
#>  $ call        : language predict.msNMix(object = object, X.0 = X.0, ignore.RE = ignore.RE, type = type)
#>  - attr(*, "class")= chr "predict.msDS"