Function for Fitting Multi-Species Abundance GLMs

The function msAbund fits multi-species abundance GLMs.

Usage

msAbund(formula, data, inits, priors, tuning, 
        n.batch, batch.length, accept.rate = 0.43, family = 'Poisson',
  n.omp.threads = 1, verbose = TRUE, n.report = 100, 
        n.burn = round(.10 * n.batch * batch.length), n.thin = 1, n.chains = 1,
        ...)

Arguments

formula

a symbolic description of the model to be fit for the model using R's model syntax. Only right-hand side of formula is specified. See example below. Random intercepts and slopes are allowed using lme4 syntax (Bates et al. 2015).

data

a list containing data necessary for model fitting. Valid tags are y and covs. y is a two or three-dimensional array of observed count data. The first dimension of the array is equal to the number of species and the second dimension is equal to the number of sites. If specified as a three-dimensional array, the third dimension corresponds to replicate observations at each site (e.g., sub-samples, repeated sampling over multiple seasons). covs is a list or data frame containing the variables used in the model. Each list element is a different covariate, which can be site-level or observation-level. Site-level covariates are specified as a vector of length \(J\), while observation-level covariates are specified as a matrix or data frame with the number of rows equal to \(J\) and number of columns equal to the maximum number of replicate observations at a given site. For zero-inflated Gaussian models, the tag z is used to specify the binary component of the model and should have the same dimensions as y.

inits

a list with each tag corresponding to a parameter name. Valid tags are beta.comm, beta, tau.sq.beta, sigma.sq.mu, kappa, tau.sq. kappa is only specified if family = 'NB', tau.sq is only specified for Gaussian or zero-inflated Gaussian models, and sigma.sq.mu is only specified if random effects are included in formula. The value portion of each tag is the parameter's initial value. See priors description for definition of each parameter name. Additionally, the tag fix can be set to TRUE to fix the starting values across all chains. If fix is not specified (the default), starting values are varied randomly across chains.

priors

a list with each tag corresponding to a parameter name. Valid tags are beta.comm.normal, tau.sq.beta.ig, sigma.sq.mu, kappa.unif, tau.sq.ig. Community-level (beta.comm) regression coefficients are assumed to follow a normal distribution. The hyperparameters of the normal distribution are passed as a list of length two with the first and second elements corresponding to the mean and variance of the normal distribution, which are each specified as vectors of length equal to the number of coefficients to be estimated or of length one if priors are the same for all coefficients. If not specified, prior means are set to 0 and prior variances to 100. Community-level variance parameters (tau.sq.beta) are assumed to follow an inverse Gamma distribution. The hyperparameters of the inverse gamma distribution are passed as a list of length two with the first and second elements corresponding to the shape and scale parameters, which are each specified as vectors of length equal to the number of coefficients to be estimated or a single value if priors are the same for all parameters. If not specified, prior shape and scale parameters are set to 0.1. sigma.sq.mu are the random effect variances random effects, respectively, and are assumed to follow an inverse Gamma distribution. The hyperparameters of the inverse-Gamma distribution are passed as a list of length two with first and second elements corresponding to the shape and scale parameters, respectively, which are each specified as vectors of length equal to the number of random intercepts or of length one if priors are the same for all random effect variances. kappa is the negative binomial dispersion parameter for each species and is assumed to follow a uniform distribution. The hyperparameters of the uniform distribution are passed as a list of length two with first and second elements corresponding to the lower and upper bounds of the uniform distribution, respectively, which are each specified as vectors of length equal to the number of species or of length one if priors are the same for all species-specific dispersion parameters. tau.sq is the species-specific residual variance for Gaussian (or zero-inflated Gaussian) models, and it is assigned an inverse-Gamma prior. The hyperparameters of the inverse-Gamma are passed as a list of length two, with the first and second element corresponding to the shape and scale parameters, respectively, which are each specified as vectors of length equal to the number of species or a single value if priors are the same for all species.

tuning

a list with each tag corresponding to a parameter name, whose whose value defines the initial variance of the adaptive sampler. Valid tags are beta, beta.star (the abundance random effect values), and kappa. See Roberts and Rosenthal (2009) for details. Note that no tuning is necessary for Gaussian or zero-inflated Gaussian models.

n.batch

the number of MCMC batches in each chain to run for the Adaptive MCMC sampler. See Roberts and Rosenthal (2009) for details.

batch.length

the length of each MCMC batch to run for the Adaptive MCMC sampler. See Roberts and Rosenthal (2009) for details.

accept.rate

target acceptance rate for Adaptive MCMC. Defaul is 0.43. See Roberts and Rosenthal (2009) for details.

family

the distribution to use for the latent abundance process. Currently supports 'NB' (negative binomial), 'Poisson', 'Gaussian', and 'zi-Gaussian'.

n.omp.threads

a positive integer indicating the number of threads to use for SMP parallel processing. The package must be compiled for OpenMP support. For most Intel-based machines, we recommend setting n.omp.threads up to the number of hyperthreaded cores. Note, n.omp.threads > 1 might not work on some systems.

verbose

if TRUE, messages about data preparation, model specification, and progress of the sampler are printed to the screen. Otherwise, no messages are printed.

n.report

the interval to report Metropolis sampler acceptance and MCMC progress. Note this is specified in terms of batches and not overall samples for spatial models.

n.burn

the number of samples out of the total n.samples to discard as burn-in for each chain. By default, the first 10% of samples is discarded.

n.thin

the thinning interval for collection of MCMC samples. The thinning occurs after the n.burn samples are discarded. Default value is set to 1.

n.chains

the number of chains to run in sequence.

...

currently no additional arguments

References

Roberts, G.O. and Rosenthal J.S. (2009) Examples of adaptive MCMC. Journal of Computational and Graphical Statistics, 18(2):349-367.

Bates, Douglas, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01 .

Author

Jeffrey W. Doser doserjef@msu.edu,
Andrew O. Finley finleya@msu.edu,

Value

An object of class msAbund that is a list comprised of:

beta.comm.samples

a coda object of posterior samples for the community level regression coefficients.

tau.sq.beta.samples

a coda object of posterior samples for the abundance community variance parameters.

beta.samples

a coda object of posterior samples for the species level abundance regression coefficients.

kappa.samples

a coda object of posterior samples for the species level abundance dispersion parameters. Only included when family = 'NB'.

tau.sq.samples

a coda object of posterior samples for the Gaussian residual variance parameter. Only included when family = 'Gaussian' or family = 'zi-Gaussian'.

y.rep.samples

a three or four-dimensional array of posterior samples for the fitted (replicate) values for each species with dimensions corresponding to MCMC sample, species, site, and replicate.

mu.samples

a three or four-dimensional array of posterior samples for the expected abundance values for each species with dimensions corresponding to MCMC samples, species, site, and replicate.

sigma.sq.mu.samples

a coda object of posterior samples for variances of random effects included in the abundance portion of the model. Only included if random effects are specified in abund.formula.

beta.star.samples

a coda object of posterior samples for the abundance random effects. Only included if random effects are specified in abund.formula.

like.samples

a three-dimensional array of posterior samples for the likelihood value associated with each site and species. Used for calculating WAIC.

rhat

a list of Gelman-Rubin diagnostic values for some of the model parameters.

ESS

a list of effective sample sizes for some of the model parameters.

run.time

MCMC sampler execution time reported using proc.time().

The return object will include additional objects used for subsequent prediction and/or model fit evaluation.

Examples

set.seed(408)
J.x <- 8
J.y <- 8
J <- J.x * J.y
n.rep <- sample(3, size = J, replace = TRUE)
n.sp <- 6
# Community-level covariate effects
beta.mean <- c(-2, 0.5)
p.abund <- length(beta.mean)
tau.sq.beta <- c(0.2, 1.2)
# Random effects (two random intercepts)
mu.RE <- list(levels = c(10, 15),
         sigma.sq.mu = c(0.43, 0.5))
# Draw species-level effects from community means.
beta <- matrix(NA, nrow = n.sp, ncol = p.abund)
for (i in 1:p.abund) {
  beta[, i] <- rnorm(n.sp, beta.mean[i], sqrt(tau.sq.beta[i]))
}
sp <- FALSE
kappa <- runif(n.sp, 0.1, 1)

dat <- simMsAbund(J.x = J.x, J.y = J.y, n.rep = n.rep, n.sp = n.sp, beta = beta,
            mu.RE = mu.RE, sp = sp, kappa = kappa, family = 'NB')

y <- dat$y
X <- dat$X
X.re <- dat$X.re

# Package all data into a list
covs <- list(int = X[, , 1],
       abund.cov.1 = X[, , 2],
       abund.factor.1 = X.re[, , 1],
       abund.factor.2 = X.re[, , 2])
data.list <- list(y = y,
      covs = covs)
prior.list <- list(beta.comm.normal = list(mean = 0, var = 100),
                   kappa.unif = list(a = 0, b = 10),
                   tau.sq.beta.ig = list(a = .1, b = .1))
inits.list <- list(beta.comm = 0,
       beta = 0,
       kappa = 0.5,
       tau.sq.beta = 1)
tuning.list <- list(kappa = 0.3, beta = 0.1, beta.star = 0.1)

# Small
n.batch <- 2
batch.length <- 25
n.burn <- 20
n.thin <- 1
n.chains <- 1

out <- msAbund(formula = ~ abund.cov.1 + (1 | abund.factor.1) + (1 | abund.factor.2),
           data = data.list,
           n.batch = n.batch,
           inits = inits.list,
           priors = prior.list,
           tuning = tuning.list,
           batch.length = batch.length,
           n.omp.threads = 3,
           verbose = TRUE,
           n.report = 1,
           n.burn = n.burn,
           n.thin = n.thin,
           n.chains = n.chains)
#> ----------------------------------------
#> 	Preparing to run the model
#> ----------------------------------------
#> No prior specified for sigma.sq.mu.ig.
#> Setting prior shape to 0.1 and prior scale to 0.1
#> sigma.sq.mu is not specified in initial values.
#> Setting initial values to random values between 0.05 and 1
#> ----------------------------------------
#> 	Model description
#> ----------------------------------------
#> Multi-species Poisson Abundance model fit with 64 sites and 6 species.
#> 
#> Samples per Chain: 50 
#> Burn-in: 20 
#> Thinning Rate: 1 
#> Number of Chains: 1 
#> Total Posterior Samples: 30 
#> 
#> Source compiled with OpenMP support and model fit using 3 thread(s).
#> 
#> Adaptive Metropolis with target acceptance rate: 43.0
#> ----------------------------------------
#> 	Chain 1
#> ----------------------------------------
#> Sampling ... 
#> Batch: 1 of 2, 50.00%
#> 	Species	Parameter	Acceptance	Tuning
#> 	1	beta[1]		68.0		0.10202
#> 	2	beta[1]		76.0		0.10202
#> 	3	beta[1]		80.0		0.10202
#> 	4	beta[1]		80.0		0.10202
#> 	5	beta[1]		48.0		0.10202
#> 	6	beta[1]		56.0		0.10000
#> -------------------------------------------------
#> Batch: 2 of 2, 100.00%
summary(out)
#> 
#> Call:
#> msAbund(formula = ~abund.cov.1 + (1 | abund.factor.1) + (1 | 
#>     abund.factor.2), data = data.list, inits = inits.list, priors = prior.list, 
#>     tuning = tuning.list, n.batch = n.batch, batch.length = batch.length, 
#>     n.omp.threads = 3, verbose = TRUE, n.report = 1, n.burn = n.burn, 
#>     n.thin = n.thin, n.chains = n.chains)
#> 
#> Samples per Chain: 50
#> Burn-in: 20
#> Thinning Rate: 1
#> Number of Chains: 1
#> Total Posterior Samples: 30
#> Run Time (min): 0.0013
#> 
#> ----------------------------------------
#> 	Community Level
#> ----------------------------------------
#> Abundance Means (log scale): 
#>                Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept) -0.9646 0.2550 -1.4283 -0.9675 -0.5500   NA   3
#> abund.cov.1 -0.0591 0.4984 -1.2478  0.0041  0.6328   NA  30
#> 
#> Abundance Variances (log scale): 
#>               Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept) 0.1698 0.1513 0.0392 0.1185 0.5531   NA  30
#> abund.cov.1 1.3626 1.6447 0.2610 0.6937 6.6265   NA   9
#> 
#> Abundance Random Effect Variances (log scale): 
#>                              Mean     SD   2.5%    50%  97.5% Rhat ESS
#> (Intercept)-abund.factor.1 0.4058 0.1089 0.2576 0.3991 0.6420   NA   3
#> (Intercept)-abund.factor.2 0.5588 0.0957 0.4365 0.5345 0.7422   NA   7
#> 
#> ----------------------------------------
#> 	Species Level
#> ----------------------------------------
#> Abundance (log scale): 
#>                    Mean     SD    2.5%     50%   97.5% Rhat ESS
#> (Intercept)-sp1 -1.0197 0.2252 -1.2552 -1.0873 -0.6356   NA   2
#> (Intercept)-sp2 -0.8911 0.0721 -1.0273 -0.9079 -0.7569   NA   7
#> (Intercept)-sp3 -1.0199 0.3226 -1.5078 -0.9727 -0.4759   NA   2
#> (Intercept)-sp4 -1.3827 0.3429 -1.7203 -1.4787 -0.6879   NA   2
#> (Intercept)-sp5 -1.1025 0.2429 -1.4599 -1.0569 -0.7740   NA   2
#> (Intercept)-sp6 -0.5708 0.3488 -1.2013 -0.5004 -0.1836   NA   1
#> abund.cov.1-sp1 -0.8751 0.1529 -1.0765 -0.9125 -0.5943   NA   3
#> abund.cov.1-sp2  0.6237 0.1242  0.3534  0.6513  0.7553   NA   6
#> abund.cov.1-sp3 -0.4646 0.0730 -0.6461 -0.4562 -0.3703   NA   6
#> abund.cov.1-sp4 -0.4174 0.1311 -0.6721 -0.4335 -0.1694   NA   7
#> abund.cov.1-sp5  0.9042 0.2054  0.5648  0.9224  1.1521   NA   1
#> abund.cov.1-sp6  0.4857 0.2051  0.1512  0.5265  0.7224   NA   2